refactor: replace original host module by host_ghislaine
